Moved deltas into their own package to prevent cyclic imports
