+++ delta/delta_suite_test.go
@@ -1,0 +1,1 @@
+package delta_test
