--- rpc/marshal_test.go
@@ -1,1 +1,0 @@
-package rpc_test
