+++ smpc/delta.go
@@ -1,21 +1,18 @@
 package smpc
 
 import (
-	"bytes"
 	"sync"
 	"time"
 
-	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/jbenet/go-base58"
-	"github.com/republicprotocol/republic-go/order"
+	"github.com/republicprotocol/republic-go/delta"
 	"github.com/republicprotocol/republic-go/shamir"
 	"github.com/republicprotocol/republic-go/stackint"
 )
 
 // OrderTuplesToDeltaFragments reads OrderTuples from an input channel and
 // uses them to build DeltaFragments.
-func OrderTuplesToDeltaFragments(done <-chan struct{}, orderTuples <-chan OrderTuple, bufferLimit int) <-chan DeltaFragment {
-	deltaFragments := make(chan DeltaFragment, bufferLimit)
+func OrderTuplesToDeltaFragments(done <-chan struct{}, orderTuples <-chan OrderTuple, bufferLimit int) <-chan delta.Fragment {
+	deltaFragments := make(chan delta.Fragment, bufferLimit)
 
 	go func() {
 		defer close(deltaFragments)
@@ -28,7 +25,7 @@ func OrderTuplesToDeltaFragments(done <-chan struct{}, orderTuples <-chan OrderT
 				if !ok {
 					return
 				}
-				deltaFragment := NewDeltaFragment(orderTuple.BuyOrderFragment, orderTuple.SellOrderFragment, &Prime)
+				deltaFragment := delta.NewDeltaFragment(orderTuple.BuyOrderFragment, orderTuple.SellOrderFragment, &Prime)
 				select {
 				case <-done:
 					return
@@ -44,8 +41,8 @@ func OrderTuplesToDeltaFragments(done <-chan struct{}, orderTuples <-chan OrderT
 // BuildDeltas by reading DeltaFragments from an input channel and using a
 // SharedDeltaBuilder to store them. Deltas can be built from the
 // SharedDeltaBuilder after a threshold of DeltaFragments has been reached.
-func BuildDeltas(done <-chan struct{}, deltaFragments <-chan DeltaFragment, sharedDeltaBuilder *SharedDeltaBuilder, bufferLimit int) <-chan Delta {
-	deltas := make(chan Delta, bufferLimit)
+func BuildDeltas(done <-chan struct{}, deltaFragments <-chan delta.Fragment, sharedDeltaBuilder *SharedDeltaBuilder, bufferLimit int) <-chan delta.Delta {
+	deltas := make(chan delta.Delta, bufferLimit)
 
 	// Insert DeltaFragments into the SharedDeltaBuilder
 	go func() {
@@ -67,7 +64,7 @@ func BuildDeltas(done <-chan struct{}, deltaFragments <-chan DeltaFragment, shar
 	go func() {
 		defer close(deltas)
 
-		buffer := make([]Delta, bufferLimit)
+		buffer := make([]delta.Delta, bufferLimit)
 		tick := time.NewTicker(time.Millisecond)
 		defer tick.Stop()
 
@@ -100,10 +97,10 @@ type SharedDeltaBuilder struct {
 	maxVolumeSharesCache shamir.Shares
 
 	mu                     *sync.Mutex
-	deltas                 map[string]Delta
-	deltaFragments         map[string]DeltaFragment
-	deltasToDeltaFragments map[string]DeltaFragments
-	deltasQueue            Deltas
+	deltas                 map[string]delta.Delta
+	deltaFragments         map[string]delta.Fragment
+	deltasToDeltaFragments map[string]delta.Fragments
+	deltasQueue            delta.Deltas
 }
 
 func NewSharedDeltaBuilder(k int64, prime stackint.Int1024) SharedDeltaBuilder {
@@ -116,14 +113,14 @@ func NewSharedDeltaBuilder(k int64, prime stackint.Int1024) SharedDeltaBuilder {
 		minVolumeSharesCache:   make(shamir.Shares, k),
 		maxVolumeSharesCache:   make(shamir.Shares, k),
 		mu:                     new(sync.Mutex),
-		deltas:                 map[string]Delta{},
-		deltaFragments:         map[string]DeltaFragment{},
-		deltasToDeltaFragments: map[string]DeltaFragments{},
-		deltasQueue:            Deltas{},
+		deltas:                 map[string]delta.Delta{},
+		deltaFragments:         map[string]delta.Fragment{},
+		deltasToDeltaFragments: map[string]delta.Fragments{},
+		deltasQueue:            delta.Deltas{},
 	}
 }
 
-func (builder *SharedDeltaBuilder) InsertDeltaFragment(deltaFragment DeltaFragment) {
+func (builder *SharedDeltaBuilder) InsertDeltaFragment(deltaFragment delta.Fragment) {
 	builder.mu.Lock()
 	defer builder.mu.Unlock()
 
@@ -142,13 +139,13 @@ func (builder *SharedDeltaBuilder) InsertDeltaFragment(deltaFragment DeltaFragme
 		builder.deltasToDeltaFragments[string(deltaFragment.DeltaID)] =
 			append(builder.deltasToDeltaFragments[string(deltaFragment.DeltaID)], deltaFragment)
 	} else {
-		builder.deltasToDeltaFragments[string(deltaFragment.DeltaID)] = DeltaFragments{deltaFragment}
+		builder.deltasToDeltaFragments[string(deltaFragment.DeltaID)] = delta.Fragments{deltaFragment}
 	}
 
 	// Build the Delta
 	deltaFragments := builder.deltasToDeltaFragments[string(deltaFragment.DeltaID)]
 	if int64(len(deltaFragments)) >= builder.k {
-		if !IsCompatible(deltaFragments) {
+		if !delta.IsCompatible(deltaFragments) {
 			return
 		}
 
@@ -160,7 +157,7 @@ func (builder *SharedDeltaBuilder) InsertDeltaFragment(deltaFragment DeltaFragme
 			builder.minVolumeSharesCache[i] = deltaFragments[i].MinVolumeShare
 		}
 
-		delta := NewDeltaFromShares(
+		delta := delta.NewDeltaFromShares(
 			deltaFragments[0].BuyOrderID,
 			deltaFragments[0].SellOrderID,
 			builder.fstCodeSharesCache,
@@ -175,7 +172,7 @@ func (builder *SharedDeltaBuilder) InsertDeltaFragment(deltaFragment DeltaFragme
 	}
 }
 
-func (builder *SharedDeltaBuilder) Deltas(deltas Deltas) int {
+func (builder *SharedDeltaBuilder) Deltas(deltas delta.Deltas) int {
 	builder.mu.Lock()
 	defer builder.mu.Unlock()
 
@@ -193,191 +190,12 @@ func (builder *SharedDeltaBuilder) Deltas(deltas Deltas) int {
 	return n
 }
 
-func (builder *SharedDeltaBuilder) hasDelta(deltaID DeltaID) bool {
+func (builder *SharedDeltaBuilder) hasDelta(deltaID delta.ID) bool {
 	_, ok := builder.deltas[string(deltaID)]
 	return ok
 }
 
-func (builder *SharedDeltaBuilder) hasDeltaFragment(deltaFragmentID DeltaFragmentID) bool {
+func (builder *SharedDeltaBuilder) hasDeltaFragment(deltaFragmentID delta.FragmentID) bool {
 	_, ok := builder.deltaFragments[string(deltaFragmentID)]
 	return ok
-}
-
-// A DeltaID is the Keccak256 hash of the order IDs that were used to compute
-// the associated Delta.
-type DeltaID []byte
-
-// Equal returns an equality check between two DeltaIDs.
-func (id DeltaID) Equal(other DeltaID) bool {
-	return bytes.Equal(id, other)
-}
-
-// String returns a DeltaID as a Base58 encoded string.
-func (id DeltaID) String() string {
-	return base58.Encode(id)
-}
-
-type Deltas []Delta
-
-type Delta struct {
-	ID          DeltaID
-	BuyOrderID  order.ID
-	SellOrderID order.ID
-	FstCode     stackint.Int1024
-	SndCode     stackint.Int1024
-	Price       stackint.Int1024
-	MaxVolume   stackint.Int1024
-	MinVolume   stackint.Int1024
-}
-
-func NewDeltaFromShares(buyOrderID, sellOrderID order.ID, fstCodeShares, sndCodeShares, priceShares, maxVolumeShares, minVolumeShares shamir.Shares, k int64, prime stackint.Int1024) Delta {
-	// Join the Shares into a Result.
-	delta := Delta{
-		BuyOrderID:  buyOrderID,
-		SellOrderID: sellOrderID,
-	}
-	delta.FstCode = shamir.Join(&prime, fstCodeShares)
-	delta.SndCode = shamir.Join(&prime, sndCodeShares)
-	delta.Price = shamir.Join(&prime, priceShares)
-	delta.MaxVolume = shamir.Join(&prime, maxVolumeShares)
-	delta.MinVolume = shamir.Join(&prime, minVolumeShares)
-
-	// Compute the ResultID and return the Result.
-	delta.ID = DeltaID(crypto.Keccak256(delta.BuyOrderID[:], delta.SellOrderID[:]))
-	return delta
-}
-
-func (delta *Delta) IsMatch(prime stackint.Int1024) bool {
-	zero := stackint.Zero()
-	two := stackint.Two()
-	zeroThreshold := prime.Div(&two)
-	if delta.FstCode.Cmp(&zero) != 0 {
-		return false
-	}
-	if delta.SndCode.Cmp(&zero) != 0 {
-		return false
-	}
-	if delta.Price.Cmp(&zeroThreshold) == 1 {
-		return false
-	}
-	if delta.MaxVolume.Cmp(&zeroThreshold) == 1 {
-		return false
-	}
-	if delta.MinVolume.Cmp(&zeroThreshold) == 1 {
-		return false
-	}
-	return true
-}
-
-// A DeltaFragmentID is the Keccak256 hash of the order IDs that were used to
-// compute the associated DeltaFragment.
-type DeltaFragmentID []byte
-
-// Equal returns an equality check between two DeltaFragmentIDs.
-func (id DeltaFragmentID) Equal(other DeltaFragmentID) bool {
-	return bytes.Equal(id, other)
-}
-
-// String returns a DeltaFragmentID as a Base58 encoded string.
-func (id DeltaFragmentID) String() string {
-	return base58.Encode(id)
-}
-
-type DeltaFragments []DeltaFragment
-
-// A DeltaFragment is a secret share of a Final. Is is performing a
-// computation over two OrderFragments.
-type DeltaFragment struct {
-	ID                  DeltaFragmentID
-	DeltaID             DeltaID
-	BuyOrderID          order.ID
-	SellOrderID         order.ID
-	BuyOrderFragmentID  order.FragmentID
-	SellOrderFragmentID order.FragmentID
-
-	FstCodeShare   shamir.Share
-	SndCodeShare   shamir.Share
-	PriceShare     shamir.Share
-	MaxVolumeShare shamir.Share
-	MinVolumeShare shamir.Share
-}
-
-func NewDeltaFragment(left, right *order.Fragment, prime *stackint.Int1024) DeltaFragment {
-	var buyOrderFragment, sellOrderFragment *order.Fragment
-	if left.OrderParity == order.ParityBuy {
-		buyOrderFragment = left
-		sellOrderFragment = right
-	} else {
-		buyOrderFragment = right
-		sellOrderFragment = left
-	}
-
-	fstCodeShare := shamir.Share{
-		Key:   buyOrderFragment.FstCodeShare.Key,
-		Value: buyOrderFragment.FstCodeShare.Value.SubModulo(&sellOrderFragment.FstCodeShare.Value, prime),
-	}
-	sndCodeShare := shamir.Share{
-		Key:   buyOrderFragment.SndCodeShare.Key,
-		Value: buyOrderFragment.SndCodeShare.Value.SubModulo(&sellOrderFragment.SndCodeShare.Value, prime),
-	}
-	priceShare := shamir.Share{
-		Key:   buyOrderFragment.PriceShare.Key,
-		Value: buyOrderFragment.PriceShare.Value.SubModulo(&sellOrderFragment.PriceShare.Value, prime),
-	}
-	maxVolumeShare := shamir.Share{
-		Key:   buyOrderFragment.MaxVolumeShare.Key,
-		Value: buyOrderFragment.MaxVolumeShare.Value.SubModulo(&sellOrderFragment.MinVolumeShare.Value, prime),
-	}
-	minVolumeShare := shamir.Share{
-		Key:   buyOrderFragment.MinVolumeShare.Key,
-		Value: sellOrderFragment.MaxVolumeShare.Value.SubModulo(&buyOrderFragment.MinVolumeShare.Value, prime),
-	}
-
-	return DeltaFragment{
-		ID:                  DeltaFragmentID(crypto.Keccak256([]byte(buyOrderFragment.ID), []byte(sellOrderFragment.ID))),
-		DeltaID:             DeltaID(crypto.Keccak256([]byte(buyOrderFragment.OrderID), []byte(sellOrderFragment.OrderID))),
-		BuyOrderID:          buyOrderFragment.OrderID,
-		SellOrderID:         sellOrderFragment.OrderID,
-		BuyOrderFragmentID:  buyOrderFragment.ID,
-		SellOrderFragmentID: sellOrderFragment.ID,
-		FstCodeShare:        fstCodeShare,
-		SndCodeShare:        sndCodeShare,
-		PriceShare:          priceShare,
-		MaxVolumeShare:      maxVolumeShare,
-		MinVolumeShare:      minVolumeShare,
-	}
-}
-
-// Equals checks if two DeltaFragments are equal in value.
-func (deltaFragment *DeltaFragment) Equals(other *DeltaFragment) bool {
-	return deltaFragment.ID.Equal(other.ID) &&
-		deltaFragment.DeltaID.Equal(other.DeltaID) &&
-		deltaFragment.BuyOrderID.Equal(other.BuyOrderID) &&
-		deltaFragment.SellOrderID.Equal(other.SellOrderID) &&
-		deltaFragment.BuyOrderFragmentID.Equal(other.BuyOrderFragmentID) &&
-		deltaFragment.SellOrderFragmentID.Equal(other.SellOrderFragmentID) &&
-		deltaFragment.FstCodeShare.Key == other.FstCodeShare.Key &&
-		deltaFragment.FstCodeShare.Value.Cmp(&other.FstCodeShare.Value) == 0 &&
-		deltaFragment.SndCodeShare.Key == other.SndCodeShare.Key &&
-		deltaFragment.SndCodeShare.Value.Cmp(&other.SndCodeShare.Value) == 0 &&
-		deltaFragment.PriceShare.Key == other.PriceShare.Key &&
-		deltaFragment.PriceShare.Value.Cmp(&other.PriceShare.Value) == 0 &&
-		deltaFragment.MaxVolumeShare.Key == other.MaxVolumeShare.Key &&
-		deltaFragment.MaxVolumeShare.Value.Cmp(&other.MaxVolumeShare.Value) == 0 &&
-		deltaFragment.MinVolumeShare.Key == other.MinVolumeShare.Key &&
-		deltaFragment.MinVolumeShare.Value.Cmp(&other.MinVolumeShare.Value) == 0
-}
-
-// IsCompatible returns true if all DeltaFragments are fragments of the same
-// Delta, otherwise it returns false.
-func IsCompatible(deltaFragments DeltaFragments) bool {
-	if len(deltaFragments) == 0 {
-		return false
-	}
-	for i := range deltaFragments {
-		if !deltaFragments[i].DeltaID.Equal(deltaFragments[0].DeltaID) {
-			return false
-		}
-	}
-	return true
 }
